requirements_2.txt
# Core
python-dotenv==1.1.0
click==8.1.8
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1

# Airflow and Providers
apache-airflow==2.7.1
apache-airflow-providers-common-sql==1.7.1
apache-airflow-providers-ftp==3.5.1
apache-airflow-providers-http==4.5.1
apache-airflow-providers-imap==3.3.1
apache-airflow-providers-sqlite==3.4.3

# Rich version compatible with Airflow & Flask-Limiter
rich==13.3.4

# Snowflake
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.25.0

# Other dependencies (only fixed if absolutely necessary)
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
anyio==4.0.0
attrs==23.1.0
boto3==1.38.25
botocore==1.38.25
cffi==1.15.1
cryptography==45.0.3
gunicorn==21.2.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.6
jsonschema==4.19.0
packaging==25.0
pendulum==2.1.2
protobuf==4.21.12
psutil==5.9.5
pyOpenSSL==25.1.0
PyJWT==2.8.0
PyYAML==6.0.2
requests==2.32.3
setproctitle==1.3.2
six==1.16.0
SQLAlchemy==1.4.49
tenacity==8.2.3
Werkzeug==2.2.3
WTForms==3.0.1
yarl==1.9.2

# Utilities and helpers (no strict pins)
colorama
colorlog
dataclasses-json
email-validator
graphviz
greenlet
importlib-resources
lazy-object-proxy
markdown-it-py
more-itertools
packaging
prompt_toolkit
PyJWT
rich-argparse
typing_extensions
tzdata
tzlocal
typing-inspection
